[refactor] 패키지 정정, 함수명 변경
