+++ app/src/main/java/com/squirtles/musicroad/map/navigation/MapNavigation.kt
@@ -11,7 +11,7 @@ import com.squirtles.musicroad.map.MapViewModel
 import com.squirtles.musicroad.media.PlayerServiceViewModel
 import com.squirtles.musicroad.navigation.MapRoute
 import com.squirtles.musicroad.navigation.Route
-import com.squirtles.musicroad.pick.PickDetailScreen
+import com.squirtles.musicroad.detail.PickDetailScreen
 
 fun NavController.navigateMap(navOptions: NavOptions? = null) {
     navigate(Route.Map, navOptions)
